:recycle: :: 변경된 Jusicool Design System에 맞게 코드 변경
