+++ lib/presentation/community_screens/community_post_write_screen.dart
@@ -1,0 +1,116 @@
+import 'package:flutter/material.dart';
+import 'package:jusicool_design_system/src/core/theme/colors/color_palette.dart';
+import 'package:jusicool_design_system/src/core/theme/texts/typography.dart';
+import 'package:jusicool_design_system/src/ui/widgets/button/button_medium.dart';
+
+class CommunityPostWriteScreen extends StatefulWidget {
+  const CommunityPostWriteScreen({super.key});
+
+  @override
+  State<CommunityPostWriteScreen> createState() => _WritePostScreenState();
+}
+
+class _WritePostScreenState extends State<CommunityPostWriteScreen> {
+  final TextEditingController titleController = TextEditingController();
+  final TextEditingController contentController = TextEditingController();
+
+  bool get isFormValid =>
+      titleController.text.trim().isNotEmpty &&
+      contentController.text.trim().isNotEmpty;
+
+  @override
+  void initState() {
+    super.initState();
+    titleController.addListener(() => setState(() {}));
+    contentController.addListener(() => setState(() {}));
+  }
+
+  @override
+  void dispose() {
+    titleController.dispose();
+    contentController.dispose();
+    super.dispose();
+  }
+
+  void _uploadPost() {
+    final title = titleController.text.trim();
+    final content = contentController.text.trim();
+    if (title.isNotEmpty && content.isNotEmpty) {
+      Navigator.pop(context, {'title': title, 'content': content});
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: JusicoolColor.white,
+      appBar: AppBar(
+        leading: const BackButton(),
+        title: Text('글 작성', style: JusicoolTypography.subTitle),
+        centerTitle: true,
+        scrolledUnderElevation: 0,
+      ),
+      body: SingleChildScrollView(
+        padding: const EdgeInsets.all(20),
+        child: ConstrainedBox(
+          constraints: BoxConstraints(
+            minHeight:
+                MediaQuery.of(context).size.height -
+                kToolbarHeight -
+                MediaQuery.of(context).padding.top -
+                80,
+          ),
+          child: IntrinsicHeight(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                TextField(
+                  controller: titleController,
+                  decoration: InputDecoration(
+                    hintText: '제목을 입력하세요',
+                    hintStyle: JusicoolTypography.titleSmall.copyWith(
+                      color: JusicoolColor.gray200,
+                    ),
+                    border: InputBorder.none,
+                  ),
+                  style: JusicoolTypography.titleSmall,
+                  keyboardType: TextInputType.multiline,
+                  maxLines: null,
+                ),
+                const SizedBox(height: 10),
+                Expanded(
+                  child: TextField(
+                    controller: contentController,
+                    decoration: InputDecoration(
+                      hintText: '내용을 입력하세요',
+                      hintStyle: JusicoolTypography.bodyMedium.copyWith(
+                        color: JusicoolColor.gray200,
+                      ),
+                      border: InputBorder.none,
+                    ),
+                    style: JusicoolTypography.bodyMedium,
+                    keyboardType: TextInputType.multiline,
+                    maxLines: null,
+                    minLines: 10,
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+      bottomNavigationBar: SafeArea(
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
+          child: AppButtonMedium(
+            text: '올리기',
+            onPressed: isFormValid ? _uploadPost : null,
+            backgroundColor: isFormValid ? JusicoolColor.main : JusicoolColor.gray400,
+            textColor: isFormValid ? JusicoolColor.white : JusicoolColor.gray600,
+            borderColor: Colors.transparent,
+          ),
+        ),
+      ),
+    );
+  }
+}
