+++ pubspec.yaml
@@ -19,10 +19,7 @@ dependencies:
   fl_chart: ^0.64.0
   intl: ^0.20.2
   path_provider: ^2.1.5
-  jusicool_design_system:
-    git:
-      url: https://github.com/Jusicool-Ver-2-0/Jusicool-iOS-DesignSystem.git
-      ref: main
+  jusicool_design_system: ^1.0.1
 
 flutter_native_splash:
   color: "#FFFFFF"
